Rename widget.
