+++ assets/js/modules/analytics-4/components/audience-segmentation/dashboard/index.js
@@ -25,6 +25,6 @@ export { default as AudienceTooltipMessage } from './AudienceTooltipMessage';
 export { default as AudienceSelectionPanel } from './AudienceSelectionPanel';
 export { default as AudienceAreaFooter } from './AudienceAreaFooter';
 export { default as ChangeGroupsLink } from './ChangeGroupsLink';
-export { default as ConnectAnalyticsCTATileWidget } from './ConnectAnalyticsCTATileWidget';
+export { default as ConnectAnalyticsCTAWidget } from './ConnectAnalyticsCTAWidget';
 export { default as InfoNoticeWidget } from './InfoNoticeWidget';
 export { default as InfoNotice } from './InfoNotice';
