+++ assets/sass/components/analytics-4/audience-segmentation/_googlesitekit-audience-connect-analytics-cta-tile.scss
@@ -1,5 +1,5 @@
 /**
- * Audience Segmentation ConnectAnalyticsCTATileWidget styles.
+ * Audience Segmentation ConnectAnalyticsCTAWidget styles.
  *
  * Site Kit by Google, Copyright 2024 Google LLC
  *
