+++ tests/backstop/reference/google-site-kit_Modules_Analytics4_Components_AudienceSegmentation_Dashboard_ConnectAnalyticsCTAWidget_0_document_0_small.png
